requirements.txt
apispec[marshmallow]==5.2.2
apispec-webframeworks==0.5.2
boto3==1.22.10
cloudinary==1.29.0
debugpy==1.6.0
Flask==2.1.2
Flask-Caching==2.0.0
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
Flask-Migrate==3.1.0
fpdf==1.7.2
faker
healthcheck==1.3.3
marshmallow==3.15.0
passlib==1.7.4
PyMySQL==1.0.2
python-dateutil==2.8.2
python-slugify[unicode]==6.1.2
python-dotenv==0.20.0
pytest==7.1.2
pytest-flask==1.2.0
pytz==2022.1
Pillow==8.1.0
redis==4.3.1
requests==2.27.1
SQLAlchemy==1.4.36
SQLAlchemy-Utils==0.38.2
tinify==1.6.0
Werkzeug==2.1.2
waitress==2.1.1

